requirements.txt
dgl==0.9.0                        
dgllife==0.2.9                        
numpy==1.20.1           
pytorch==1.9.0           
rdkit==2023.3.3                     
scikit-learn==1.2.2                        
scipy==1.10.1                       
torch-cluster==1.5.9                        
torch-geometric==2.0.0                        
torch-scatter==2.0.8                        
torch-sparse==0.6.11                       
tqdm==4.65.0                       